notificationdrf/requirements.txt
python-dotenv==0.21.0
Django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.21.4
django-celery-beat==2.5.0
celery==5.2.7
redis==3.4.1
eventlet==0.33.3
